Add option to not delete replies to your own tweets
These 'replies' are not true replies but corrections or
simply more tweets to be read in sequence ("threads").
Hence need/must not be deleted when user chooses to delete
replies to others.

Right now user has to present their id in cmd-line option;
but internally the solution is general and uses a list, so
in future to support "spare multiple user replies" only
cmd-line parser needs to be changed. Also in future we might
read id directly from file, making it easy for user.